refactor: rename and reorganize two factor auth dialog flow
