--- app/assets/javascripts/preferences/models/index.ts
@@ -1,2 +1,0 @@
-export * from './preferences';
-export * from './two-factor-auth';
