+++ app/assets/javascripts/preferences/panes/two-factor-auth/EmailRecovery.tsx
