+++ app/assets/javascripts/preferences/panes/two-factor-auth/Verification.tsx
@@ -1,6 +1,6 @@
-import { TwoFactorActivation } from '@/preferences/models';
 import { observer } from 'mobx-react-lite';
 import { FunctionComponent } from 'preact';
+import { TwoFactorActivation } from './model';
 
 export const Verification: FunctionComponent<{
   activation: TwoFactorActivation;
