+++ app/assets/javascripts/preferences/panes/two-factor-auth/index.tsx
@@ -1,8 +1,8 @@
-import { TwoFactorAuth } from '../../models';
 import { observable } from 'mobx';
 import { observer } from 'mobx-react-lite';
 import { FunctionComponent } from 'preact';
-import { TwoFactorAuthView } from './view';
+import { TwoFactorAuth } from './model';
+import { TwoFactorAuthView } from './TwoFactorAuthView';
 
 export const TwoFactorAuthWrapper: FunctionComponent = observer(() => {
   const tfAuth = observable(new TwoFactorAuth());
