+++ app/assets/javascripts/preferences/preferences-menu.ts
@@ -1,6 +1,5 @@
 import { IconType } from '@/components/Icon';
 import { makeAutoObservable, observable } from 'mobx';
-import { TwoFactorAuth } from './two-factor-auth';
 
 const PREFERENCE_IDS = [
   'general',
